Add type inference from PyTorch types

+++ torcharrow/torcharrow/test/test_interop.py
@@ -334,7 +334,7 @@ class TestInterop(unittest.TestCase):
 
         # Reverse conversion
         p = df.to_torch()
-        df2 = tap.from_torch(p, dtype=df.dtype)
+        df2 = tap.from_torch(p)
         self.assertEqual(df.dtype, df2.dtype)
         self.assertEqual(list(df), list(df2))
